feat(suppliers): fixing layout menu billed, purchases, rma supplier, credit supplier, supplier lists

+++ src/features/pages/suppliers/billed/index.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React, { useEffect } from 'react'
+import React, { useState, useEffect } from 'react'
 import Table from "@/components/table"
 import type { TableColumnsType } from 'antd'
 import { Dropdown, Menu } from 'antd'
@@ -19,17 +19,47 @@ import { billAtom } from '@/store/SuppliersAtom'
 import { stripHTML } from '@/plugins/validators/common-rules'
 import StatusTag from '@/components/tag'
 import dayjs from 'dayjs'
+import Pagination from '@/components/pagination'
+import {
+    MoreIcon,
+    TrashIconRed,
+    AddIcon,
+    PrintIconBlack,
+    EmailBlackIcon,
+    StatusIcon,
+    DuplicateIcon,
+    WalletIcon,
+    PencilIconBlue
+} from '@public/icon'
+import Image from 'next/image'
+import ButtonIcon from '@/components/button/ButtonIcon'
+import ButtonAction from '@/components/button/ButtonAction'
+import SearchTable from '@/components/search/SearchTable'
+import ShowPageSize from '@/components/pagination/ShowPageSize'
 
 const index = ({ billData }: { billData?: any }) => {
     const { contextHolder, notifySuccess } = useNotificationAntd()
     const [data, setData] = useAtom(billAtom)
+    const [search, setSearch] = useState('')
+    const [filteredData, setFilteredData] = useState<BillType[]>([])
+    const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([]);
+    const [currentOrder, setCurrentOrder] = useState<any>(null)
+    const [currentPage, setCurrentPage] = useState(1);
+    const [pageSize, setPageSize] = useState(10);
+    const [openModalDelete, setOpenModalDelete] = useState(false)
+    const [deletedData, setDeletedData] = useState<any>(null)
 
     const handleDelete = async (id: any) => {
+        if (!deletedData) return;
         try {
             const res = await deletePurchases(id)
             if (res.success == true) {
                 notifySuccess(res.message)
                 setData(prev => prev.filter(item => item.id !== id))
+
+                //  setFilteredData(updatedData.filter(item =>
+                //         item.name.toLowerCase().includes(search.toLowerCase())
+                //     ));
             }
         } catch (error) {
             console.error(error)
@@ -65,7 +95,7 @@ const index = ({ billData }: { billData?: any }) => {
     ]
     const columns: TableColumnsType<BillType> = [
         {
-            title: 'Order Id',
+            title: 'Order ID',
             dataIndex: 'order_id',
         },
         {
@@ -75,13 +105,16 @@ const index = ({ billData }: { billData?: any }) => {
         {
             title: 'Total',
             dataIndex: 'total',
+            render: (val) => {
+                return <span>${val}</span>
+            }
         },
         {
             title: 'Status',
             dataIndex: 'status',
             render: (val) => {
-                const status = val ? stripHTML(val.toUpperCase()) : '';
-                return <StatusTag status={status} />
+                const status = val ? stripHTML(val) : '';
+                return <StatusTag status={status} type='supplier' />
             }
         },
         {
@@ -89,7 +122,7 @@ const index = ({ billData }: { billData?: any }) => {
             dataIndex: 'email_status',
             render: (val) => {
                 const status = val ? stripHTML(val) : '';
-                return status
+                return <StatusTag status={status} type='supplier' />
             }
         },
         {
@@ -162,16 +195,32 @@ const index = ({ billData }: { billData?: any }) => {
                 );
 
                 return (
-                    <Dropdown overlay={menu} trigger={['click']} >
-                        <button className="flex items-center justify-center px-2 py-1 border rounded ">
-                            Actions <MoreOutlined className="ml-1" />
-                        </button>
-                    </Dropdown >
+                    <div className="flex items-center justify-end gap-3 pe-4">
+                        <ButtonIcon
+                            color='primary'
+                            variant='filled'
+                            size="small"
+                            icon={PencilIconBlue}
+                        // onClick={() => router.push(routes.eCommerce.editAttributes(row.id))}
+                        />
+                        <ButtonIcon
+                            color='danger'
+                            variant='filled'
+                            size="small"
+                            icon={TrashIconRed}
+                            onClick={() => handleOpenModalDelete(row.id)}
+                        />
+                    </div >
                 );
             }
         },
 
     ]
+
+    const handleOpenModalDelete = (data: any) => {
+        setOpenModalDelete(true)
+        setDeletedData(data)
+    }
 
     const handleSearch = (query: string) => {
         console.log('User mencari:', query);
@@ -184,34 +233,135 @@ const index = ({ billData }: { billData?: any }) => {
         <>
             {contextHolder}
             <div className="mt-6 mx-4 mb-0">
-                <h1 className='text-xl font-bold'>
-                    Billed
-                </h1>
-                <Breadcrumb
-                    items={breadcrumb}
-                />
+                <div className='flex justify-between items-center'>
+                    <div>
+                        <h1 className='text-xl font-bold'>
+                            Billed
+                        </h1>
+                        <Breadcrumb
+                            items={breadcrumb}
+                        />
+                    </div>
+                    <Button
+                        icon={<Image
+                            src={AddIcon}
+                            alt='add-icon'
+                            width={15}
+                            height={15}
+                        />}
+                        label='Add Bill'
+                        link={routes.eCommerce.createBill}
+                    />
+                </div>
             </div>
-            <Content className="mt-6 mx-4 mb-0">
-                <div style={{ padding: 24, minHeight: 360, background: '#fff' }}>
-                    <div className='flex justify-end mb-4'>
+            <Content className="mb-0">
+                <div style={{ padding: 24, minHeight: 360 }}>
+                    <div className='flex justify-between mb-4 gap-2'>
                         <div className='flex items-center gap-2'>
-                            <SearchInput onSearch={handleSearch} />
-                            <Button
+                            <ShowPageSize
+                                pageSize={pageSize}
+                                onChange={setPageSize}
+                            />
 
-                                icon={<PlusCircleOutlined />}
-                                label='Add Bill'
-                                link={routes.eCommerce.createBill}
+                            <SearchTable
+                                value={search}
+                                onChange={(e) => setSearch(e.target.value)}
+                                onSearch={() => console.log('Searching for:', search)}
                             />
                         </div>
-
+                        {
+                            selectedRowKeys.length > 0 &&
+                            <div className='flex  gap-3'>
+                                <ButtonAction
+                                    icon={<Image
+                                        src={PrintIconBlack}
+                                        alt='print-icon'
+                                        width={15}
+                                        height={15}
+                                    />}
+                                    label='Print'
+                                    style={{ padding: '1.2rem 1.7rem' }}
+                                    onClick={() => console.log('hi')}
+                                />
+                                <ButtonAction
+                                    icon={<Image
+                                        src={EmailBlackIcon}
+                                        alt='email-icon'
+                                        width={15}
+                                        height={15}
+                                    />}
+                                    label='Email'
+                                    style={{ padding: '1.2rem 1.7rem' }}
+                                    onClick={() => console.log('hi')}
+                                />
+                                <ButtonAction
+                                    icon={<Image
+                                        src={StatusIcon}
+                                        alt='status-icon'
+                                        width={15}
+                                        height={15}
+                                    />}
+                                    label='Status'
+                                    style={{ padding: '1.2rem 1.7rem' }}
+                                    onClick={() => console.log('hi')}
+                                />
+                                <ButtonAction
+                                    icon={<Image
+                                        src={DuplicateIcon}
+                                        alt='duplicate-icon'
+                                        width={15}
+                                        height={15}
+                                    />}
+                                    label='Duplicate'
+                                    style={{ padding: '1.2rem 1.7rem' }}
+                                    onClick={() => console.log('hi')}
+                                />
+                                {
+                                    selectedRowKeys.length == 1 &&
+                                    <ButtonAction
+                                        icon={<Image
+                                            src={WalletIcon}
+                                            alt='paynow-icon'
+                                            width={15}
+                                            height={15}
+                                        />}
+                                        label='Pay Now'
+                                        style={{ padding: '1.2rem 1.7rem' }}
+                                        onClick={() => console.log('hi')}
+                                    />
+                                }
+                                <ButtonAction
+                                    label='Delete All'
+                                    icon={<Image
+                                        src={TrashIconRed}
+                                        alt='trash-icon'
+                                        width={10}
+                                        height={10}
+                                    />}
+                                    // onClick={() => setisOpenModalFilter(true)}
+                                    position='start'
+                                    style={{ padding: '1.2rem 1.7rem' }}
+                                    btnClassname='btn-delete-all'
+                                />
+                            </div>
+                        }
                     </div>
                     <Table
                         columns={columns}
                         dataSource={data}
                         withSelectableRows
+                        selectedRowKeys={selectedRowKeys}
+                        onSelectChange={setSelectedRowKeys}
+                    />
+                    <Pagination
+                        current={currentPage}
+                        total={data.length}
+                        pageSize={pageSize}
+                        onChange={(page) => setCurrentPage(page)}
                     />
                 </div>
             </Content>
+
         </>
     )
 }
